chore: apply feedback from review - part 1
